requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.4
colorama==0.4.6
dnspython==2.4.0
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
python-dotenv==1.0.0
sniffio==1.3.0
Werkzeug==2.3.6
gunicorn
